Rename device file
